Date fix for Safari
There's a known behavior in Safari that applies current OS time zone to the date object when the Date constructor (or Date.parse method) is used. Because of this serializr will produce the date of 2017/12/07 in New Zealand time zone for the value of '2017-12-06T15:37:39.903'. The suggested fix will mitigate the issue for ISO 8601 format date strings without timezones.

See the following links for more details:
https://stackoverflow.com/a/33909265
https://stackoverflow.com/a/6427318